backend/AILocationService/AILocationService/requirements.txt
fastapi
uvicorn[standard]
openai
geopy
python-dotenv
requests
python-jose[cryptography]
passlib[bcrypt]
foursquare

# Database dependencies
sqlalchemy==2.0.23
geoalchemy2==0.14.0
psycopg2-binary==2.9.9
shapely==2.0.1
